% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{abbreviate_number}
\alias{abbreviate_number}
\title{Abbreviate Number}
\usage{
abbreviate_number(number)
}
\arguments{
\item{number}{A numeric vector containing the numbers to be abbreviated.}
}
\value{
A list containing the abbreviated numbers and their corresponding suffixes.
}
\description{
This function abbreviates large numbers using a standardized notation.
}
\details{
This function takes a numeric vector and abbreviates large numbers using a standardized notation, such as "K" for thousand, "M" for million, etc. It calculates the appropriate abbreviation based on the magnitude of the input numbers.
The following abbreviations are used:
\itemize{
\item -: <1
\item : 1 to 100
\item H: Hundred (10^2)
\item K: Thousand (10^3)
\item M: Million (10^6)
\item B: Billion (10^9)
\item T: Trillion (10^12)
\item Q: Quadrillion (10^15)
\item QQ: Quintillion (10^18)
\item S: Sextillion (10^21)
\item SS: Septillion (10^24)
\item O: Octillion (10^27)
\item N: Nonillion (10^30)
\item D: Decillion (10^33)
\item UD: Undecillion (10^36)
\item DDD: Duodecillion (10^39)
\item TD: Tredecillion (10^42)
\item QD: Quattuordecillion (10^45)
\item QTD: Quindecillion (10^48)
\item SD: Sexdecillion (10^51)
\item SSD: Septendecillion (10^54)
\item OD: Octodecillion (10^57)
\item NND: Novemdecillion (10^60)
\item V: Vigintillion (10^63)
\item UV: Unvigintillion (10^66)
\item DV: Duovigintillion (10^69)
\item TV: Trevigintillion (10^72)
\item QV: Quattuorvigintillion (10^75)
\item QTV: Quinvigintillion (10^78)
\item SV: Sexvigintillion (10^81)
\item SSV: Septenvigintillion (10^84)
\item OV: Octovigintillion (10^87)
\item NNV: Novemvigintillion (10^90)
\item X: Trigintillion (10^93)
}
}
\examples{
\dontrun{
# Do not run examples because they may take a long time
abbreviate_number(c(1000, 1000000, 1000000000))
}

}
