% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{advertising}
\alias{advertising}
\title{Advertising Dataset}
\format{
A data frame with 13 rows and 4 variables:
\describe{
\item{TV}{Advertising dollars spent on TV for a single product (in thousands of dollars).}
\item{Radio}{Advertising dollars spent on Radio.}
\item{Newspaper}{Advertising dollars spent on Newspaper.}
\item{Sales}{Sales of a single product (in thousands of units).}
}
}
\source{
This dataset is a subset of the Advertising dataset from "An Introduction to Statistical Learning" (ISL).
}
\usage{
advertising
}
\description{
This dataset contains information about advertising spending and sales.
}
\examples{
\dontrun{
  head(advertising)
}

}
\keyword{datasets}
