% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{all_combination_of_flexi_vars}
\alias{all_combination_of_flexi_vars}
\title{Get all combinations of variables from a named list}
\usage{
all_combination_of_flexi_vars(flexi_vars)
}
\arguments{
\item{flexi_vars}{A named list where each element represents a variable.}
}
\value{
A list containing all combinations of variables from flexi_vars.
}
\description{
This function generates all combinations of variables (flexible variables)
from a given named list.
}
\examples{
\dontrun{
flexi_vars <- list(a = c(m = 0, n = 1, o = 2),
     b = c(m = 0.1, n = 2.1, o = 4.2), 
     c = c(m = 3.1, n = 2.7, o = 4.5))
all_combination_of_flexi_vars(flexi_vars)
}
}
