% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{compare_and_print_differences}
\alias{compare_and_print_differences}
\title{Compare and print named vectors}
\usage{
compare_and_print_differences(
  first_vector,
  second_vector,
  acceptable_difference = NULL
)
}
\arguments{
\item{first_vector}{A named vector.}

\item{second_vector}{A named vector.}

\item{acceptable_difference}{Threshold for acceptable difference.}
}
\value{
This function does not return any value. It prints the differences between the two vectors.
}
\description{
Compare two named vectors and print the differences between them.
}
\examples{
first_vector <- c(a = 1, b = 2, c = 3)
second_vector <- c(a = 1, b = 4, d = 5)
compare_and_print_differences(first_vector, second_vector, acceptable_difference = 1)

}
