% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{convert_to_named_seq}
\alias{convert_to_named_seq}
\title{Convert String to Named Sequence}
\usage{
convert_to_named_seq(char_vector, delimiter, names_vector)
}
\arguments{
\item{char_vector}{A character vector with strings to be split and converted.}

\item{delimiter}{A character string specifying the delimiter used for splitting.}

\item{names_vector}{A character vector specifying the names to be assigned to the numeric values.}
}
\value{
A list of named numeric vectors.
}
\description{
This function takes a character vector, splits each string by a specified delimiter,
converts the split parts into numeric values, and assigns user-defined names
to these numeric values.
}
\examples{
\dontrun{
char_vector <- c("1|3|2", "4|6|2")
delimiter <- "|"
names_vector <- c("start", "end", "step")
convert_to_named_seq(char_vector, delimiter, names_vector)
}
}
