% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{convert_variable_info2list}
\alias{convert_variable_info2list}
\title{Convert Variable Information to a List}
\usage{
convert_variable_info2list(variables_wt_apl_contri)
}
\arguments{
\item{variables_wt_apl_contri}{Named numeric vector representing variables with weight, application, and contribution values.}
}
\value{
A list of numeric vectors containing variable, weight, application details, and contribution values.
}
\description{
Applies a function to convert variable information with contribution values to a list.
}
\examples{
\dontrun{
# Example 1
convert_variable_info2list(c(a_2_4_2 = 9, b_2_4_1 = 6))

# Example 2
convert_variable_info2list(c(a_2_4_2 = 9))
}

}
