% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{create_file_path}
\alias{create_file_path}
\title{Create File Path}
\usage{
create_file_path(folder, file_name, relative_directory = ".")
}
\arguments{
\item{folder}{The folder in which the file is located.}

\item{file_name}{The name of the file.}

\item{relative_directory}{The relative directory path (default is ".").}
}
\value{
A character vector representing the absolute file path.
}
\description{
Create a file path based on the specified folder, file name, and relative directory.
}
\examples{
\dontrun{
create_file_path("data", "example.csv")
create_file_path("output", "result.txt", relative_directory = "project")
}
}
