% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{create_named_lists_from_vars}
\alias{create_named_lists_from_vars}
\title{Create Named Lists from Variables with Predefined Pairs}
\usage{
create_named_lists_from_vars(var_names, predefined_pairs)
}
\arguments{
\item{var_names}{A character vector of variable names.}

\item{predefined_pairs}{A list of predefined key-value pairs to be replicated.}
}
\value{
A named list where each element is a copy of the predefined pairs list.
}
\description{
Creates a list where each element is a copy of the predefined pairs list.
The output list will have the same names as the input \code{var_names} vector.
Note that each element in the output list is a direct copy of \code{predefined_pairs},
and not a list of individual key-value pairs.
}
\examples{
\dontrun{
var_names <- c("var1", "var2")
predefined_pairs <- list(setNames(1:3, c("adstock", "power", "lag")))
create_named_lists_from_vars(var_names, predefined_pairs)
}
}
