% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_transformations.R
\name{find_critical_vif}
\alias{find_critical_vif}
\title{Determine VIF Flag for Predictors}
\usage{
find_critical_vif(type, vif_threshold)
}
\arguments{
\item{type}{Character string indicating the type of the variable.
Options are 'fixed' or 'flexible'. The function applies special
logic if there is only one 'fixed' variable.}

\item{vif_threshold}{Numeric value specifying the threshold above which the
VIF flag should be triggered.}
}
\value{
A numeric value; the function returns \code{Inf} (infinite) if there is only
one 'fixed' variable, otherwise it returns the \code{vif_threshold}.
}
\description{
This function evaluates the Variance Inflation Factor (VIF) for predictors and
determines if it exceeds a specified threshold. It accounts for special cases
based on the type of the variable.
}
\details{
The function uses the following logic:
\itemize{
\item If the \code{type} is 'fixed' and there is only one such variable, it returns \code{Inf}.
This accounts for the scenario where a single fixed variable should not trigger
a flag by having an artificially high VIF.
\item In all other cases, it returns the \code{vif_threshold}.
}
}
\examples{
\dontrun{
  find_critical_vif("fixed", 5)      # Returns Inf
  find_critical_vif("flexible", 5)   # Returns 5
  find_critical_vif("fixed|fixed", 5)  # Returns 5
}

}
