% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{flatten_path}
\alias{flatten_path}
\title{Flatten hierarchical paths into a tidy data frame.}
\usage{
flatten_path(full_path)
}
\arguments{
\item{full_path}{A character vector containing hierarchical paths separated by '||'.}
}
\value{
A tidy data frame with columns representing different levels of the hierarchy.
\itemize{
\item \code{aggregated_path}: Separate rows for each aggregated path.
\item \code{individual_path}: Paths separated by '||'.
\item \code{aggregated_ids}: IDs separated by '>'.
\item \code{granular_ids}: IDs separated by '|'.
}
}
\description{
This function takes a hierarchical path structure represented as a delimited
string and flattens it into a tidy data frame format using the tidyverse
functions.
}
\examples{
\dontrun{
full_path <- "aggregated_path1||aggregated_path2>individual_path1>individual_path2|granular_id1|granular_id2"
flatten_path(full_path)
}

}
