% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{format_number}
\alias{format_number}
\title{Format Number}
\usage{
format_number(
  numeric_vector,
  notation,
  round_digit_mapping = setNames(2, "M"),
  prefix = ""
)
}
\arguments{
\item{numeric_vector}{A numeric vector containing the numbers to be formatted.}

\item{notation}{An ordered factor specifying the notation for each number in numeric_vector.}

\item{round_digit_mapping}{A named numeric vector specifying the number of digits to round each notation to (default is set to 2 for "M" notation).}

\item{prefix}{A prefix string to be added before each formatted number (default is empty).}
}
\value{
A character vector containing the formatted numbers with the specified notation and rounding.
}
\description{
This function formats numeric vectors according to specified notation and rounding rules.
}
\details{
This function formats numeric vectors based on the specified notation and rounding rules.
It rounds the numbers according to the specified digits for each notation and adds a prefix string if provided.
The 'notation' parameter should be an ordered factor with levels representing the notation for each number in 'numeric_vector'.
The 'round_digit_mapping' parameter is a named numeric vector where the names represent the notations and the values represent the number of digits to round to.
The default rounding is set to 2 digits for "M" notation.

Numbers with corresponding notation greater than the name of 'round_digit_mapping' will be rounded to 0.
Numbers with corresponding notation equal to the name of 'round_digit_mapping' will be rounded to the value of 'round_digit_mapping'.
Numbers greater than or equal to 1 with corresponding notation less than the name of 'round_digit_mapping' will be rounded to the value of 'round_digit_mapping'.
Numbers less than 1 will be converted to a significant number of 'round_digit_mapping'.
}
\examples{
\dontrun{
format_number(c(1234567, 9876543), ordered(c("M", "K")),
 round_digit_mapping = c(M = 2), prefix = "$")
}

}
