% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_arrows}
\alias{generate_arrows}
\title{Generate arrows based on values relative to a reference point}
\usage{
generate_arrows(
  x,
  reference_point = 0,
  size = 40,
  shrinkage_factor = 0.5,
  positive_color = "#00FF00",
  negative_color = "red"
)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{reference_point}{The reference point relative to which arrows are generated. Default is 0.}

\item{size}{The size of the arrows. Default is 40.}

\item{shrinkage_factor}{A value controlling the shrinkage of arrow width based on the magnitude of the input values.
A smaller shrinkage_factor will result in less relative difference in arrow sizes for values closer to the reference_point.
Default is 0.5.}

\item{positive_color}{The color of arrows representing values greater than or equal to the reference point. Default is "#00FF00" (green).}

\item{negative_color}{The color of arrows representing values less than the reference point. Default is "red".}
}
\value{
A character vector containing HTML representations of arrows.
}
\description{
This function generates arrows representing the direction and magnitude of values relative to a reference point.
Arrows are HTML entities rendered based on the sign and magnitude of the input values. The size of the arrows is
determined by the absolute value of the input values raised to a power, allowing for non-linear scaling.
The closer the value to zero, the less relative difference there will be between the sizes of the arrows.
}
\examples{
\dontrun{
x <- c(1, 2, 3, 0, -1, -2, -3)
arrows <- generate_arrows(x, reference_point = 0, size = 40, shrinkage_factor = 0.5)
print(arrows)
}

}
