% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{generate_circles}
\alias{generate_circles}
\title{Generate circles with static color}
\usage{
generate_circles(x, max_radius = 10, min_radius = 5, circle_color = "blue")
}
\arguments{
\item{x}{A numeric vector of values.}

\item{max_radius}{The maximum radius of the circles. Default is 10.}

\item{min_radius}{The minimum radius of the circles. Default is 5.}

\item{circle_color}{The color of the circles. Default is "blue".}
}
\value{
A character vector containing HTML representations of circles.
}
\description{
This function generates circles with a static color based on the values provided in the input vector.
The size of the circles is determined by scaling the values linearly between a maximum and minimum radius.
The color of all circles remains constant and is specified by the circle_color parameter.
}
\examples{
\dontrun{
x <- c(1, 2, 3, 4, 5)
circles <- generate_circles(x, max_radius = 20, min_radius = 10, circle_color = "red")
print(circles)
}

}
