% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_transformations.R
\name{generate_model_decomposition}
\alias{generate_model_decomposition}
\title{Generate Model Decomposition}
\usage{
generate_model_decomposition(
  dep_info,
  indep_info,
  modeling_df,
  dep_info_is_weight_coefficient = FALSE,
  indep_info_is_weight_coefficient = TRUE,
  apl_delimiter = "_",
  delimiter = "_",
  var_agg_delimiter = "|"
)
}
\arguments{
\item{dep_info}{A list specifying the dependent variable(s) or combinations of dependent variables along with their transformation parameters. Each element of the list should be a numeric value representing the weight coefficient for the corresponding dependent variable(s). The names of dependent variables should include the adstock, power, and lag parameters required for variable transformation, along with variable names and components, separated by the delimiter.}

\item{indep_info}{A list specifying the independent variable(s) or combinations of independent variables along with their transformation parameters. Each element of the list should be a numeric value representing the weight coefficient for the corresponding independent variable(s). The names of independent variables should include the adstock, power, and lag parameters required for variable transformation, along with variable names and components, separated by the delimiter.}

\item{modeling_df}{The data frame containing the modeling data.}

\item{dep_info_is_weight_coefficient}{Logical, indicating whether the dependent variable information represents weight coefficients.}

\item{indep_info_is_weight_coefficient}{Logical, indicating whether the independent variable information represents weight coefficients.}

\item{apl_delimiter}{The delimiter used for separating adstock, power, lag parameters from variable names and components.}

\item{delimiter}{The delimiter used for separating variable components.}

\item{var_agg_delimiter}{The delimiter used for variable aggregation.}
}
\value{
A list containing the dependent and its decomposition of its components.
}
\description{
This function decomposes a model into its components, including dependent and independent variables, based on provided transformation parameters. It allows for the decomposition of complex models into simpler components for analysis or interpretation.
}
\examples{
\dontrun{
df <- data.frame(
  dep_var1 = c(1, 2, 3),
  dep_var2 = c(4, 5, 6),
  indep_var1 = c(7, 8, 9),
  indep_var2 = c(10, 11, 12),
  indep_var3 = c(13, 14, 15)
)
generate_model_decomposition(list(setNames(1,"dep_var1_0_1_0")), list(setNames(c(1:3),
  c("indep_var1_0_1_0","indep_var2_0_1_0","indep_var3_0_1_0"))), df)
generate_model_decomposition(list(setNames(1,"dep_var1|dep_var2_0_1_0")),
  list(setNames(c(1:2),c("indep_var1|indep_var2_0_1_0","indep_var3_0_1_0"))), df)
generate_model_decomposition(list(setNames(1:2,c("dep_var1_0_1_0","dep_var2_0_1_0"))),
 list(setNames(c(1:3),c("indep_var1_0_1_0","indep_var2_0_1_0","indep_var3_0_1_0"))), df)
generate_model_decomposition(list(setNames(1,"dep_var1_0_1_0"),setNames(2,"dep_var2_0_1_0")),
 list(setNames(c(1:3),c("indep_var1_0_1_0","indep_var2_0_1_0","indep_var3_0_1_0")),
 setNames(c(1:3),c("indep_var1_0_1_0","indep_var2_0_1_0","indep_var3_0_1_0"))), df)
}
}
