% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_model_flags_plot}
\alias{generate_model_flags_plot}
\title{Generate a Venn diagram plot summarizing model flags}
\usage{
generate_model_flags_plot(pvalue_flag, sign_flag, vif_flag)
}
\arguments{
\item{pvalue_flag}{A logical vector indicating the presence or absence of p-value flags for each observation.}

\item{sign_flag}{A logical vector indicating the presence or absence of sign flags for each observation.}

\item{vif_flag}{A logical vector indicating the presence or absence of VIF flags for each observation.}
}
\value{
A ggplot object representing the Venn diagram plot.
}
\description{
This function generates a Venn diagram plot summarizing the model flags provided.Each flag represent to given model.
}
\examples{
\dontrun{
# Generate sample data
pvalue_flag <- sample(c(T,F), 100, replace = TRUE)
sign_flag <- sample(c(T,F), 100, replace = TRUE)
vif_flag <- sample(c(T,F), 100, replace = TRUE)

# Call the function with sample data
generate_model_flags_plot(pvalue_flag, sign_flag, vif_flag)
}

}
