% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{generate_reverse_target_models}
\alias{generate_reverse_target_models}
\title{Generate Reverse Target Models}
\usage{
generate_reverse_target_models(
  final_contri,
  collated_models,
  dep_contri_ranges_wt_model_ids,
  acceptable_lower_diff = 1e-04,
  acceptable_upper_diff = 1e-04,
  round_digits = 10,
  verbose = F
)
}
\arguments{
\item{final_contri}{A dataframe with the following columns:
"variable", "adstock", "power", "lag", "dep_contri_exp".
This dataframe represents the top hierarchy (root) of dependent attributes used to trace the model.
It must comply with the feasible conditions specified in dep_contri_ranges_wt_model_ids.}

\item{collated_models}{A dataframe containing information on successful models at each model ID.
The columns in collated_models include:
"stage_id", "version_id", "dependent_id", "dep_variable", "dep_adstock", "dep_power", "dep_lag", "model_id", "loop_id",
"indep_variable", "indep_adstock", "indep_power", "indep_lag", "contri_perc", "stage_version_dep_mdl_loop_stage".
Rows must be sorted by stage IDs, where lower row numbers represent granular or leaf models, and higher row numbers indicate top hierarchy in models.
All model IDs must be present in collated_models.}

\item{dep_contri_ranges_wt_model_ids}{A dataframe containing model IDs, dependent variables, and acceptable contributions (contri_min & contri_max) to achieve benchmark APL and contribution.
This dataframe includes the contribution mix and maximum contribution at each model stage necessary to achieve model benchmark.
Rows must be sorted by stage IDs, where lower row numbers represent granular or leaf models, and higher row numbers indicate top hierarchy in models.
The file is generated as part of the process for reverse modeling.}

\item{acceptable_lower_diff}{Acceptable lower difference for contribution matching.}

\item{acceptable_upper_diff}{Acceptable upper difference for contribution matching.}

\item{round_digits}{Number of digits to which adstock, power, and lag should be rounded to resolve issues of approximation while merging.}

\item{verbose}{If TRUE, prints information on intermediate steps.}
}
\value{
A dataframe with all combinations from model collation including intermediate contributions.
}
\description{
This function generates reverse target models based on provided final contributions and model summaries.
It takes input dataframes containing final contributions (final_contri), collated models (collated_models),
and successful model information (dep_contri_ranges_wt_model_ids), and performs matching and accumulation to generate the models.
The resulting dataframe contains all combinations of final contributions from collated models along with intermediate contributions.
}
