% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{generate_reverse_target_smry}
\alias{generate_reverse_target_smry}
\title{Summarizes target model at stage and variable level.}
\usage{
generate_reverse_target_smry(stage_id_target_model)
}
\arguments{
\item{stage_id_target_model}{A dataframe containing information about the target model.}
}
\value{
A dataframe summarizing the target model at stage and variable level.
}
\description{
This function takes a dataframe containing information about the target model
at different stages and variables. It calculates summary statistics such as
model hierarchy, number of unique models (apl), average contribution (contri_avg),
and contribution deviation (contri_dev).
}
