% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_dep_vars_combinations}
\alias{get_dep_vars_combinations}
\title{Generates combinations of dependent variables based on provided possible values.}
\usage{
get_dep_vars_combinations(dep_vars, possible_values)
}
\arguments{
\item{dep_vars}{List of dependent variable names.}

\item{possible_values}{List containing data frames with possible values for each dependent variable.}
}
\value{
A list of named vectors representing all combinations of dependent variables.
}
\description{
Generates combinations of dependent variables based on provided possible values.
}
\examples{
\dontrun{
# Example 1
dep_var <- c("a", "b")
possible_values <- list(c(a_1_9_2 = 1, b_1_3_4 = 2, a_1_4_4 = 10, b_1_5_3 = 3))
get_dep_vars_combinations(dep_var, possible_values)

# Example 2
dep_var <- c("a")
possible_values <- list(c(a_1_4_4 = 10, a_1_4_5 = 10, a_1_1_4 = 10))
get_dep_vars_combinations(dep_var, possible_values)
}

}
