% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{name_and_truncate_list}
\alias{name_and_truncate_list}
\title{Name and Truncate List}
\usage{
name_and_truncate_list(list_to_name, var_names)
}
\arguments{
\item{list_to_name}{A list to be named.}

\item{var_names}{A vector of names to be assigned to the list elements.}
}
\value{
A named list.
}
\description{
Takes a list and a vector of variable names, truncates the list
to the length of the variable names (if necessary), and sets the names
of the list elements to those variable names. If the list is shorter
than the names vector, the function truncates the names vector to the
length of the list.
}
\examples{
\dontrun{
list_to_name <- list(1, 2, 3, 4)
var_names <- c("var1", "var2")
name_and_truncate_list(list_to_name, var_names)
}
}
