% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{perform_optimization}
\alias{perform_optimization}
\title{Perform Optimization}
\usage{
perform_optimization(
  constr_type,
  constr_value,
  alpha,
  beta,
  initial_spend,
  lower_bound_spend,
  upper_bound_spend
)
}
\arguments{
\item{constr_type}{Type of constraint: "budget" or "goal".}

\item{constr_value}{Numeric value for the constraint.}

\item{alpha}{Numeric vector representing model coefficients.}

\item{beta}{Numeric vector representing model powers.}

\item{initial_spend}{Initial spending values for optimization.}

\item{lower_bound_spend}{Lower bound for spending optimization.}

\item{upper_bound_spend}{Upper bound for spending optimization.}
}
\value{
Numeric vector representing the optimized solution.
}
\description{
This function conducts optimization based on specified constraints and model parameters.
}
\details{
The function employs the NLOPT_LD_SLSQP algorithm for optimization.
For "goal" constraints, the optimization maximizes the sum of spending.
For "budget" constraints, the optimization minimizes the negative sum of spending.
}
\examples{
\dontrun{
# Example usage
constr_type <- "Goal"
constr_value <- 5
alpha <- c(1, 2)
beta <- c(0.3, 0.4)
initial_spend <- c(2, 2)
lower_bound_spend <- c(-1000, -10000)
upper_bound_spend <- c(10000, 10000)
perform_optimization(constr_type, constr_value, alpha, beta, initial_spend, lower_bound_spend,
 upper_bound_spend)
}

}
