% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_contri_dist}
\alias{plot_contri_dist}
\title{Plot Contribution Percentage Distribution}
\usage{
plot_contri_dist(adstock, power, contri_perc)
}
\arguments{
\item{adstock}{A vector containing adstock values.}

\item{power}{A vector containing power values.}

\item{contri_perc}{A vector containing contribution percentages.}
}
\value{
A ggplot object displaying the distribution of contribution percentages.
}
\description{
Plots the distribution of contribution percentages across different combinations
of adstock and power values.
}
\examples{
\dontrun{
#Generate dummy data
adstock_values <- seq(0.1, 0.9, by = 0.1)
power_values <- seq(0.1, 0.9, by = 0.1)
contri_df <- expand.grid(adstock = adstock_values, power = power_values)
# Increase rows to 1000
contri_df <- bind_rows(replicate(10, contri_df, simplify = FALSE))
# Add contribution percentages around 100 with some variability
contri_df$contri_perc <- rnorm(810, mean = 100, sd = 20)
# Plot the contribution percentage distribution
plot_contri_dist(contri_df$adstock, contri_df$power, contri_df$contri_perc)
}

}
