% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{recursive_model_summary}
\alias{recursive_model_summary}
\title{Summarize Recursive Model Coefficients}
\usage{
recursive_model_summary(
  model_dependent,
  model_smry,
  model_coef,
  round_digits = 10,
  ids_separator = "|"
)
}
\arguments{
\item{model_dependent}{Dataset containing information about the dependent variable in the model.}

\item{model_smry}{Summary dataset related to the model.}

\item{model_coef}{Dataset containing coefficients of the model.}

\item{round_digits}{Number of digits to round up numeric values to. It is important to specify
this parameter due to potential approximation issues when merging datasets with different precision.}

\item{ids_separator}{used to concatenate ids for model}
}
\value{
A summarized dataset containing information about the model coefficients.
}
\description{
This function summarizes the coefficients of a recursive model, joining relevant
datasets and performing necessary data transformations. It takes dependent variable information,
model summary, and model coefficient datasets, and consolidates them using unique identifiers
to create a comprehensive summary file with both dependent and independent information.
}
