% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{replace_values_pairwise}
\alias{replace_values_pairwise}
\title{Replace Values Pairwise}
\usage{
replace_values_pairwise(
  input_vector,
  conditions_replacements,
  replacement_values
)
}
\arguments{
\item{input_vector}{An input vector with values to be replaced.}

\item{conditions_replacements}{A vector of conditions to be checked.}

\item{replacement_values}{A vector of replacement values.}
}
\value{
A vector with replaced values.
}
\description{
Replaces each value in an input vector that matches a condition from a list
of conditions with the corresponding value from a list of replacement values.
Error if lengths of conditions and replacements are not equal.
}
\examples{
\dontrun{
input_vector <- c(1, 2, 3, 4, 5)
conditions_replacements <- c(2, 4)
replacement_values <- c(20, 40)
replace_values_pairwise(input_vector, conditions_replacements, replacement_values)
}
}
