% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_transformations.R
\name{replicate_and_extend_dep_ids}
\alias{replicate_and_extend_dep_ids}
\title{Replicate and Extend Dependent IDs with Factors}
\usage{
replicate_and_extend_dep_ids(data, ids_with_factors, cols_to_apply_factor)
}
\arguments{
\item{data}{The original dataset to be replicated and extended.}

\item{ids_with_factors}{A named list where each element corresponds to a dependent ID,
and the values are vectors of factors to be applied.}

\item{cols_to_apply_factor}{A character vector specifying the columns in the dataset
on which the factors should be applied.}
}
\value{
A replicated and extended dataset with modified values in specified columns.
}
\description{
This function replicates and extends a dataset by applying factors to specific columns
based on the provided dependent IDs and factors.
}
\examples{
\dontrun{
# Set a seed for reproducibility
set.seed(123)

# Number of rows in the dataset
num_rows <- 100

# Generate data
data <- data.frame(
  dependent_id = 1:num_rows,
  age = sample(22:45, num_rows, replace = TRUE),
  income = sample(40000:100000, num_rows, replace = TRUE),
  expenses = sample(1500:5000, num_rows, replace = TRUE),
  savings = sample(18000:75000, num_rows, replace = TRUE)
)

# Display the first few rows of the dataset
head(data)

# Define factors to apply for specific dependent IDs and columns
ids_with_factors <- list("2" = c(2, 4), "3" = c(3, 4, 5))
cols_to_apply_factor <- c("income", "savings")

# Replicate and extend the dataset
replicate_and_extend_dep_ids(data, ids_with_factors, cols_to_apply_factor)
}

}
