% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{scd_type_2_update}
\alias{scd_type_2_update}
\title{Slowly Changing Dimension Type 2 Update Function}
\usage{
scd_type_2_update(
  base_data,
  current_data,
  changes_ids,
  row_ids,
  compare_cols,
  create_date = "start_date",
  update_date = "end_date"
)
}
\arguments{
\item{base_data}{A dataframe representing the base data to be updated.}

\item{current_data}{A dataframe representing the current data with potential updates.}

\item{changes_ids}{A vector of column names at which base data will always be obsolete.
This is a higher hierarchy of data in which changes are traced and old data will be obsolete.}

\item{row_ids}{A vector of column names with primary keys.}

\item{compare_cols}{A vector of column names used for comparing record values.}

\item{create_date}{The name of the column representing creation date in the base data.}

\item{update_date}{The name of the column representing update date in the base data.}
}
\value{
A dataframe that is the result of applying SCD2 methodology on the base data
with the updates from the current data.
}
\description{
This function updates a base dataset with changes from a current dataset using the
Slowly Changing Dimension Type 2 (SCD2) methodology. It handles additions of new
records and updates to existing records based on specified identification columns
and comparison columns. It assumes that the base data should not contain records
ending with '\emph{base', and new data should not contain 'start_date' and 'end_date'.
Additionally, it only compares numeric values in \code{compare_cols} and excludes any
columns starting with 'source}' in \code{current_data}.
}
\examples{
\dontrun{
  base_data <- data.frame(
    model_id = c(1, 1, 2, 3),
    version_id = c(1, 2, 3, 4),
    name = c("Alice", "Bob", "Charlie", "David"),
    value = c(100, 200, 300, 400),
    start_date = as.POSIXct(c("2023-01-01", "2023-01-05", "2023-01-05", "2023-01-05")),
    end_date = as.POSIXct(NA)
  )
  current_data <- data.frame(
    model_id = c(1),
    version_id = c(2),
    name = c("Alice"),
    value = c(300)
  )
  changes_ids <- c("model_id")
  row_ids <- c("model_id", "version_id")
  compare_cols <- "value"
  scd_type_2_update(base_data, current_data, changes_ids, row_ids, compare_cols)
}

}
