% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{split_and_convert_to_numeric}
\alias{split_and_convert_to_numeric}
\title{Split String and Convert to Numeric}
\usage{
split_and_convert_to_numeric(char_vector, delimiter)
}
\arguments{
\item{char_vector}{A character vector with strings to be split and converted to numeric.}

\item{delimiter}{A character string specifying the delimiter used for splitting.}
}
\value{
A list of numeric vectors.
}
\description{
Splits each string in a character vector by a specified delimiter and converts
the resulting parts to numeric vectors. Returns a list of these numeric vectors.
}
\examples{
\dontrun{
char_vector <- c("1,2,3", "4,5,6")
delimiter <- ","
split_and_convert_to_numeric(char_vector, delimiter)
}
}
