% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{subset_feasible_target_models}
\alias{subset_feasible_target_models}
\title{Subset Feasible Target Models}
\usage{
subset_feasible_target_models(
  model_summary,
  model_dependent,
  target_model,
  current_stage_id,
  independent_variables,
  contri_min_relaxation_factor = 1,
  contri_max_relaxation_factor = 1,
  target_contri_round = 5,
  verbose = FALSE
)
}
\arguments{
\item{model_summary}{A data frame summarizing the model, with columns including:
\itemize{
\item \code{stage_id}: Stage ID
\item \code{version_id}: Version ID
\item \code{dependent_id}: Dependent variable ID
\item \code{model_id}: Model ID
\item \code{loop_id}: Loop ID
\item \code{dependent_sum}: Sum of the dependent variable
\item \code{fixed_contri_perc}: Fixed contribution percentage
\item Additional columns for each independent variable, indicating adstock, power, lag, contribution, and contribution percentage.
}}

\item{model_dependent}{A data frame with details for the dependent model, linked to \code{dependent_id} in \code{model_summary}. Columns include:
\itemize{
\item \code{stage_id}: Stage ID
\item \code{version_id}: Version ID
\item \code{dependent_id}: Dependent variable ID
\item \code{variable}: Independent variable name
\item \code{adstock}: Adstock effect
\item \code{power}: Power of the model
\item \code{lag}: Lag effect
\item \code{sum_of_dependent_series}: Sum of the dependent variable series
}}

\item{target_model}{A data frame with target model details to check against the feasible space. Includes:
\itemize{
\item \code{stage_id}: Stage ID
\item \code{variable}: Independent variable name
\item \code{adstock}: Adstock effect
\item \code{power}: Power of the model
\item \code{lag}: Lag effect
\item \code{contri_min}: Minimum contribution value
\item \code{contri_max}: Maximum contribution value
}}

\item{current_stage_id}{Current stage ID.}

\item{independent_variables}{A character vector of independent variables in the model.}

\item{contri_min_relaxation_factor}{A numeric factor for relaxing the minimum contribution.}

\item{contri_max_relaxation_factor}{A numeric factor for relaxing the maximum contribution.}

\item{target_contri_round}{Numeric precision for rounding contribution values.}

\item{verbose}{Logical, whether to print processing details.}
}
\value{
A data frame with updated target models, showing minimized and maximized contributions for feasible models.
}
\description{
Filters target models based on feasibility criteria and updates the target model. This is useful for identifying and working with models that fit within a specified contribution range.
}
