% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{trim_list_for_dependent_variables}
\alias{trim_list_for_dependent_variables}
\title{Find Final Dependent Variables}
\usage{
trim_list_for_dependent_variables(
  list_of_dep_var_scope,
  dep_vars,
  var_agg_delimiter = "|",
  apl_delimiter = "_",
  delimiter = "_"
)
}
\arguments{
\item{list_of_dep_var_scope}{A list containing variable scopes.}

\item{dep_vars}{Character vector of dependent variable names to be matched.}

\item{var_agg_delimiter}{A character specifying the variable aggregation delimiter. default value is "|".}

\item{apl_delimiter}{A character specifying the APL delimiter. default value is "_".}

\item{delimiter}{A character specifying the variable APL delimiter. default value is "_".}
}
\value{
A subset of list_of_dep_var_scope containing the final dependent variables that match the specified dep_vars.
}
\description{
This function identifies and returns the final dependent variables from a list of variable scopes based on the provided dependent variable names. It is used to create final dependent named vector for model
}
\examples{
\dontrun{
list_of_vars <- list(a_1_9_2 = 1, b_1_3_4 = 2, c_3_5_2=9, a_1_4_4=10,b_1_5_3=3)
dep_vars <- c("a", "b")
find_final_dependent_vars(list_of_vars, dep_vars)
}

}
